final feature added

+++ resturant_project_mobile/lib/features/favorite_screen/data/model/favorite_model.dart
@@ -127,7 +127,7 @@ class Restaurant {
             )
           : [],
 
-      rating: json["rating"] ?? 0,
+      rating: (json["rating"] ?? 0).toDouble(),
 
       delivery: json["delivery"] ?? false,
